Moving new method from PM to app
